Minor update Fix bug  	modified:   del_all_qsub_jobs.sh

+++ bsub.sh
@@ -12,8 +12,14 @@ sub_()
 
    echo $1 $2 $3
 
-   rm ${PL}_${XS}.*.${ENERGY}.${TARG}.log.gz
-   gzip ${PL}_${XS}.*.${ENERGY}.${TARG}.log 
+   ls ${PL}_${XS}.*.${ENERGY}.${TARG}.log.gz &> /dev/null
+   if [ $? = 0 ] ; then
+      rm ${PL}_${XS}.*.${ENERGY}.${TARG}.log.gz
+   fi
+   ls ${PL}_${XS}.*.${ENERGY}.${TARG}.log &> /dev/null
+   if [ $? = 0 ] ; then
+      gzip ${PL}_${XS}.*.${ENERGY}.${TARG}.log 
+   fi
 
    n=1
    nmax=16
@@ -24,7 +30,7 @@ sub_()
 #      ./SATIF_BNL ${TARG}.${ENERGY}.in ${PL} ${XS} ${n} > ${PL}_${XS}.${n}job.${ENERGY}.${TARG}.log &
        qsub -v WORKDIR=$PWD,TARG=$TARG,ENERGY=$ENERGY,PL=$PL,XS=$XS,n=$n,OUTPUTFILE=${PL}_${XS}.${n}job.${ENERGY}.${TARG}.log ./torque.sh 
        #sh ./torque.sh ${TARG}.${ENERGY}.in ${PL} ${XS} ${n} ${PL}_${XS}.${n}job.${ENERGY}.${TARG}.log
-       sleep 2
+       sleep 1
 
       n=`expr $n + 1` 
    done
@@ -40,6 +46,10 @@ targ_energy_()
    sub_ $PL Concrete 24GeV
 }
 
+ls batch_job.* &> /dev/null
+if [ $? = 0 ] ; then
+   rm batch_job.*
+fi
 #targ_energy_ Shielding
 targ_energy_ QGSP_BERT
 #sub_ QGSP_BERT
